made training header dynamic with content
