+++ src/pages/client/TrainingPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 
 import { RouteComponentProps } from "react-router-dom";
 
-import ScaledAgileHeader from "../../components/Header/ScaledAgileHeader";
+import TrainingHeader from "../../components/Header/TrainingHeader";
 import ContentContainer from "../../components/ContentContainer/ContentContainer";
 import ClassList from "../../components/TrainingClasses/ClassList";
 
@@ -12,16 +12,24 @@ interface Props extends RouteComponentProps {}
 
 export type ClassType = "LeSS" | "scaled-agile" | undefined;
 
-const TrainingPage: React.FC<Props> = ({ ...props }: Props) => {
+const TrainingPage: React.FC<Props> = ({ ...props }) => {
   const { location } = useHistory();
 
   return (
     <>
-      <ScaledAgileHeader
-        title="Scaled Agile 5.0 Certification Training"
-        description="Master the enterprise at scale"
-        type="SA"
-      />
+      {location.pathname === "/training/scaled-agile" ? (
+        <TrainingHeader
+          title="Scaled Agile 5.0 Certification Training"
+          description="Master the enterprise at scale"
+          type="SA"
+        />
+      ) : (
+        <TrainingHeader
+          title="LeSS Certification Training"
+          description="Do more with LeSS"
+          type="LeSS"
+        />
+      )}
       <ContentContainer>
         <ClassList classType={location.pathname} />
       </ContentContainer>
